recommendations-be/requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
Django==4.1.2
django-cors-headers==3.13.0
django-stubs==1.12.0
django-stubs-ext==0.5.0
djangorestframework==3.14.0
flake8==2.3.0
isort==5.10.1
mccabe==0.7.0
mypy==0.982
mypy-extensions==0.4.3
pathspec==0.10.1
pep8==1.7.1
platformdirs==2.5.2
psycopg2-binary==2.9.4
pyflakes==2.5.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
types-pytz==2022.5.0.0
types-PyYAML==6.0.12.1
typing_extensions==4.4.0
